moving and unifying Detoxify to metrics from /de and /en
